+++ 06_3D空間実験室/app/assets/rgbdDecode.fragment-DPucXRWu.js
@@ -1,0 +1,5 @@
+import{m as r}from"./index-CnFcGove.js";import"./helperFunctions-B2iYpzAC.js";const e="rgbdDecodePixelShader",o=`varying vec2 vUV;uniform sampler2D textureSampler;
+#include<helperFunctions>
+#define CUSTOM_FRAGMENT_DEFINITIONS
+void main(void) 
+{gl_FragColor=vec4(fromRGBD(texture2D(textureSampler,vUV)),1.0);}`;r.ShadersStore[e]||(r.ShadersStore[e]=o);const d={name:e,shader:o};export{d as rgbdDecodePixelShader};